Extract write_gba_file to lib.rs
